Made black holes stop mauling you posthumously
Enforced respawning rules, fixed warnings, and swapped out the text when
you die and can't respawn.

+++ Code/Game/TheGame.cpp
@@ -281,7 +281,7 @@ void TheGame::PressStart(NamedProperties&)
 
     BeginTransitioning();
     m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelStarWipe")->m_texture);
-    m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+    m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
     m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
     AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
 }
@@ -356,7 +356,7 @@ void TheGame::UpdatePlayerJoin(float)
 
             BeginTransitioning();
             m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelStarWipe")->m_texture);
-            m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+            m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
             m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
             AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
             return;
@@ -437,7 +437,7 @@ void TheGame::UpdateAssemblyGetReady(float deltaSeconds)
 
             BeginTransitioning();
             m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-            m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+            m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
             m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
             AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
             return;
@@ -520,7 +520,7 @@ void TheGame::UpdateAssemblyPlaying(float deltaSeconds)
 
         BeginTransitioning();
         m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-        m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+        m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
         m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
         AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
     }
@@ -627,7 +627,7 @@ void TheGame::UpdateAssemblyResults(float deltaSeconds)
         
         BeginTransitioning();
         m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-        m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+        m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
         m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
         AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
     }
@@ -679,7 +679,7 @@ void TheGame::UpdateMinigameGetReady(float deltaSeconds)
         
             BeginTransitioning();
             m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-            m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+            m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
             m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
             AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
             return;
@@ -749,7 +749,7 @@ void TheGame::UpdateMinigamePlaying(float deltaSeconds)
     
         BeginTransitioning();
         m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-        m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+        m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
         m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
         AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
     }
@@ -839,7 +839,7 @@ void TheGame::UpdateMinigameResults(float deltaSeconds)
         
             BeginTransitioning();
             m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-            m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+            m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
             m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
             AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
         }
@@ -854,7 +854,7 @@ void TheGame::UpdateMinigameResults(float deltaSeconds)
 
             BeginTransitioning();
             m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-            m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+            m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
             m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
             AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
         }
@@ -912,7 +912,7 @@ void TheGame::UpdateGameOver(float )
         
         BeginTransitioning();
         m_transitionFBOEffect->SetNormalTexture(ResourceDatabase::instance->GetSpriteResource("PixelWipeRight")->m_texture);
-        m_transitionFBOEffect->SetFloatUniform("gEffectTime", GetCurrentTimeSeconds());
+        m_transitionFBOEffect->SetFloatUniform("gEffectTime", (float)GetCurrentTimeSeconds());
         m_transitionFBOEffect->SetVec4Uniform("gWipeColor", RGBA::KINDA_GRAY.ToVec4());
         AudioSystem::instance->PlaySound(SFX_UI_ADVANCE);
     }
